refactor (PanelsWrapper): move a few functions in utils/panel
