+++ src/utils/panel.tsx
@@ -29,3 +29,20 @@ export function isItemContentValid(itemData: Item): boolean {
   if (itemData.content.length === 0) return false
   return true
 }
+
+export function getContentType(value: string): string {
+  let type = value.split('type=')[1]
+  type = type.charAt(0).toUpperCase() + type.slice(1); // convert the first letter to upper case
+  return type ?? 'missing'
+  // when no string stays after type=, then the value is missing
+}
+
+
+export function getContentTypes(content: Content[]): string[] {
+  const types: string[] = content.map((item) => {
+    if ('type' in item) return getContentType(item.type)
+    return 'missing'
+  })
+
+  return types
+}
